// vulpes_Front-End-adriana-dev2/src/Interfaces/cadernos/caderno3.tsx
export interface IPage3 {
  id_cara_pecu: number ;
  epecie: string | undefined;
  grupo_homoge: string | undefined;
  classe_etaria: string | undefined;
  modo_producao: string | undefined;
  naturais: string | undefined;
  normais: string | undefined;
  finalidade_producao: string | undefined;
  observacoes: string | undefined;
  id_rosto: number | undefined;
  last_update: string | undefined;
  create_date: string | undefined;
  uuid: string | undefined;
}